+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* [BREAKING] Updated `isSerializable`. Additionaly to previous behaviour now it returns true for primitives
+  (boolean/Boolean, number/Number, string/String), both instances and constructor functions. Returns true for null value.
 * [BREAKING] Removed `@Serialize.Skip()` decorator which was used to "erase" property serializers inherited from parent class.
   This practice breaks inheritance. E.g.: We have `B extends A`, B "shadows" some property serializers from A. If we'll try to
   serialize an instance of B and then deserialize it as an instance of A, we'll have an error because of undefined props.